lambda_src/create_event/requirements.txt
PyJWT==2.10.1
boto3==1.38.32
cryptography>=3.4.7